refactor: unifies gamestate class for client and server
