+++ setup.py
@@ -1,0 +1,12 @@
+import setuptools
+
+setuptools.setup(
+    name='verfuchst',
+    version='0.0.1',
+    url='https://github.com/yths/yths.verfuchst.git',
+    author='@yths',
+    author_email='@yths',
+    description='A game for four players.',
+    packages=setuptools.find_packages(),
+    install_requires=[],
+)
